+++ js/functions.js
@@ -1,9 +1,10 @@
 // Задание 1
 
 const checkLength = (string, length) => string.length <= length;
-console.log(checkLength('проверяемая строка', 20)); // true
-console.log(checkLength('проверяемая строка', 18)); // true
-console.log(checkLength('проверяемая строка', 10)); // false
+checkLength();
+// console.log(checkLength('проверяемая строка', 20)); // true
+// console.log(checkLength('проверяемая строка', 18)); // true
+// console.log(checkLength('проверяемая строка', 10)); // false
 
 // Задание 2
 
@@ -19,11 +20,11 @@ const checkPalindrom = (string) => {
   }
   return false;
 };
-
-console.log(checkPalindrom('топот')); // true
-console.log(checkPalindrom('ДовОд')); // true
-console.log(checkPalindrom('Кекс')); // false
-console.log(checkPalindrom('Лёша на полке клопа нашёл')); // true
+checkPalindrom();
+// console.log(checkPalindrom('топот')); // true
+// console.log(checkPalindrom('ДовОд')); // true
+// console.log(checkPalindrom('Кекс')); // false
+// console.log(checkPalindrom('Лёша на полке клопа нашёл')); // true
 
 // Задание 3
 
@@ -37,6 +38,6 @@ const parseNumber = (string) => {
   }
   return stringNumber;
 };
-
-console.log(parseNumber('3222 какой - то левый текст')); // 3222
-console.log(parseNumber('3222 какой 777 то левый 34 текст')); // 322277734
+parseNumber();
+// console.log(parseNumber('3222 какой - то левый текст')); // 3222
+// console.log(parseNumber('3222 какой 777 то левый 34 текст')); // 322277734
