+++ js/until.js
@@ -1,0 +1,10 @@
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => getRandomInteger(elements.MIN, elements.MAX);
+
+export { getRandomInteger, getRandomArrayElement };
